+++ Editor/CASEditorSettings.cs
@@ -29,6 +29,7 @@ namespace CAS.UEditor
         public bool updateGradlePluginVersion = true;
 
         public bool generateIOSDeepLinksForPromo = true;
+        public bool bitcodeIOSEnabled = false;
         public string attributionReportEndpoint = CASEditorUtils.attributionReportEndPoint;
 
         /// <summary>