+++ Editor/CASInitSettignsInspector.cs
@@ -46,6 +46,7 @@ namespace CAS.UEditor
         private SerializedProperty mostPopularCountryOfUsersProp;
         private SerializedProperty attributionReportEndpointProp;
         private SerializedProperty userTrackingUsageDescriptionProp;
+        private SerializedProperty bitcodeIOSEnabledProp;
         #endregion
 
         #region Utility fields
@@ -132,6 +133,7 @@ namespace CAS.UEditor
 
             mostPopularCountryOfUsersProp = editorSettingsObj.FindProperty( "mostPopularCountryOfUsers" );
             attributionReportEndpointProp = editorSettingsObj.FindProperty( "attributionReportEndpoint" );
+            bitcodeIOSEnabledProp = editorSettingsObj.FindProperty( "bitcodeIOSEnabled" );
 
             userTrackingUsageDescriptionProp = editorSettingsObj.FindProperty( "userTrackingUsageDescription" );
 
@@ -392,6 +394,10 @@ namespace CAS.UEditor
             }
             else
             {
+                bitcodeIOSEnabledProp.boolValue = EditorGUILayout.ToggleLeft(
+                    "Bitcode enabled",
+                    bitcodeIOSEnabledProp.boolValue );
+
                 EditorGUILayout.BeginHorizontal();
                 var reportEndpointEnabled = attributionReportEndpointProp.stringValue.Length > 0;
                 if (reportEndpointEnabled != EditorGUILayout.ToggleLeft(